Issue #709. Moving regression test to its own test file.
