+++ prusti-tests/tests/verify/pass/arrays/repeat-expr-initialization.rs
@@ -4,23 +4,6 @@
 //   initializers like [1, 2, 3]
 // - bool as element type
 
-extern crate prusti_contracts;
-use prusti_contracts::*;
-
-#[derive(Clone, Copy)]
-pub struct A(usize);
-
-#[derive(Clone, Copy)]
-pub struct B([A; 16]);
-
-impl B {
-    #[pure]
-    /// Initialization of an array holding ADTs.
-    pub fn new() -> Self {
-        Self([A(0); 16])
-    }
-}
-
 fn main() {
     let a = [false; 3];
